part2/mongodb_consumer_mqtt/requirements.txt
paho-mqtt==1.6.1
pymongo==4.3.3